requirements.txt
streamlit==1.32.0
boto3==1.38.17
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
